refactor: extract shared post creation logic and fix collaborator invitation acceptance
# Conflicts:
#	migrations/20260401000005_add_post_likes.sql
#	migrations/20260401000006_add_guest_comments.sql
#	migrations/20260423199010_add_post_images.sql
#	migrations/20260423199011_add_password_reset_tokens.sql
#	migrations/20260423199020_multiple_locations_per_brand.sql
#	migrations/20260423199021_add_org_units.sql
