+++ Makefile
@@ -2,7 +2,6 @@
 .PHONY: test test-cover docs docs-guest docs-business docs-admin-auth
 .PHONY: generate generate-guest-business-client clean
 .PHONY: goose-up goose-down goose-status goose-create
-.PHONY: docker-build
 
 COUNT ?= 1
 MIGRATIONS_DIR := migrations
@@ -46,17 +45,17 @@ s3-up:
 	bash scripts/bootstrap.sh
 
 s3-ui:
-	docker compose -f build/compose.yaml up -d garage_webui
-	@echo "web_ui: http://localhost:4309"
+	docker compose -f docker/compose.yaml up -d garage_webui
+	@echo "web_ui: http://localhost:3909"
 
 goose-up:
-	GOOSE_DRIVER=postgres GOOSE_DBSTRING="$(DB_DSN)" goose -dir $(MIGRATIONS_DIR) up
+	goose -dir $(MIGRATIONS_DIR) postgres "$(DB_DSN)" up
 
 goose-down:
-	GOOSE_DRIVER=postgres GOOSE_DBSTRING="$(DB_DSN)" goose -dir $(MIGRATIONS_DIR) down
+	goose -dir $(MIGRATIONS_DIR) postgres "$(DB_DSN)" down
 
 goose-status:
-	GOOSE_DRIVER=postgres GOOSE_DBSTRING="$(DB_DSN)" goose -dir $(MIGRATIONS_DIR) status
+	goose -dir $(MIGRATIONS_DIR) postgres "$(DB_DSN)" status
 
 goose-create:
 	@if [ -z "$(name)" ]; then \
@@ -95,9 +94,3 @@ clean:
 	@echo "cleaning generated files..."
 	rm -rf docs/api
 	rm -rf internal/guest/gateway/business/client
-
-docker-build:
-	docker build -t guest-api -f build/Dockerfile.guest .
-	docker build -t business-api -f build/Dockerfile.business .
-	docker build -t admin-auth-api -f build/Dockerfile.admin .
-	docker build -t migrator -f build/Dockerfile.migrator .
